Update how external locations behave
