+++ rotkehlchen/db/schema.py
@@ -630,7 +630,7 @@ CREATE TABLE IF NOT EXISTS balancer_events (
 );
 """
 
-DB_CREATE_IMPORTED_LOCATIONS = """
+DB_CREATE_CONNECTED_LOCATIONS = """
 CREATE TABLE IF NOT EXISTS connected_location (
     location INTEGER PRIMARY KEY NOT NULL,
     FOREIGN KEY (location) REFERENCES location(seq)
@@ -698,7 +698,7 @@ BEGIN TRANSACTION;
 {DB_CREATE_GITCOIN_TX_TYPE}
 {DB_CREATE_GITCOIN_GRANT_METADATA}
 {DB_CREATE_NFTS}
-{DB_CREATE_IMPORTED_LOCATIONS}
+{DB_CREATE_CONNECTED_LOCATIONS}
 COMMIT;
 PRAGMA foreign_keys=on;
 """
