+++ rotkehlchen/tests/db/test_db.py
@@ -2,13 +2,11 @@ import logging
 import os
 import time
 from copy import deepcopy
-from pathlib import Path
 from shutil import copyfile
 from unittest.mock import patch
 
 
 import pytest
-import requests
 
 from rotkehlchen.accounting.ledger_actions import LedgerActionType
 from rotkehlchen.accounting.structures import ActionType, BalanceType
@@ -82,8 +80,6 @@ from rotkehlchen.user_messages import MessagesAggregator
 from rotkehlchen.utils.misc import ts_now
 from rotkehlchen.utils.serialization import rlk_jsondumps
 
-from rotkehlchen.tests.utils.api import api_url_for
-
 TABLES_AT_INIT = [
     'assets',
     'aave_events',
@@ -128,6 +124,7 @@ TABLES_AT_INIT = [
     'gitcoin_tx_type',
     'gitcoin_grant_metadata',
     'nfts',
+    'connected_location',
 ]
 
 
@@ -1330,79 +1327,3 @@ def test_binance_pairs(user_data_dir):
     db.set_binance_pairs('binance', [], Location.BINANCE)
     query = db.get_binance_pairs('binance', Location.BINANCE)
     assert query == []
-
-
-def test_known_locations(user_data_dir, rotkehlchen_api_server):
-    """Test that locations imported in different places are correctly stored in database"""
-
-    # Import from nexo
-    dir_path = Path(__file__).resolve().parent.parent
-    filepath = dir_path / 'data' / 'nexo.csv'
-
-    json_data = {'source': 'nexo', 'file': str(filepath)}
-    requests.put(
-        api_url_for(
-            rotkehlchen_api_server,
-            'dataimportresource',
-        ), json=json_data,
-    )
-
-    # Add transactions from blockfi
-    filepath = dir_path / 'data' / 'blockfi-transactions.csv'
-
-    json_data = {'source': 'blockfi-transactions', 'file': str(filepath)}
-    requests.put(
-        api_url_for(
-            rotkehlchen_api_server,
-            'dataimportresource',
-        ), json=json_data,
-    )
-
-    # Add cryptocom
-    filepath = dir_path / 'data' / 'cryptocom_trades_list.csv'
-    json_data = {'source': 'cryptocom', 'file': str(filepath)}
-    requests.put(
-        api_url_for(
-            rotkehlchen_api_server,
-            'dataimportresource',
-        ), json=json_data,
-    )
-
-    # Add cointracking multiple exchanges
-    filepath = dir_path / 'data' / 'cointracking_trades_list.csv'
-    files = {'file': open(filepath, 'rb')}
-    requests.post(
-        api_url_for(
-            rotkehlchen_api_server,
-            'dataimportresource',
-        ),
-        files=files,
-        data={'source': 'cointracking.info'},
-    )
-
-    # Add multiple entries for same exchange + connected exchange
-    msg_aggregator = MessagesAggregator()
-    db = DBHandler(user_data_dir, '123', msg_aggregator, None)
-    kraken_api_key1 = ApiKey('kraken_api_key')
-    kraken_api_secret1 = ApiSecret(b'kraken_api_secret')
-    kraken_api_key2 = ApiKey('kraken_api_key2')
-    kraken_api_secret2 = ApiSecret(b'kraken_api_secret2')
-    binance_api_key = ApiKey('binance_api_key')
-    binance_api_secret = ApiSecret(b'binance_api_secret')
-
-    # add mock kraken and binance
-    db.add_exchange('kraken1', Location.KRAKEN, kraken_api_key1, kraken_api_secret1)
-    db.add_exchange('kraken2', Location.KRAKEN, kraken_api_key2, kraken_api_secret2)
-    db.add_exchange('binance', Location.BINANCE, binance_api_key, binance_api_secret)
-
-    expected_locations = {
-        Location.KRAKEN,
-        Location.BINANCE,
-        Location.BLOCKFI,
-        Location.NEXO,
-        Location.CRYPTOCOM,
-        Location.POLONIEX,
-        Location.COINBASE,
-    }
-
-    assert set(db.get_connected_locations()) == expected_locations
